Rename to reflect the trunk version merged in.

+++ escript/src/Taipan.cpp
@@ -335,9 +335,9 @@ Taipan::dump_stats() {
 
   assert(totalElements >= 0);
 
-  float elMb=statTable->allocated_elements*8.0/1048576;
-  float deelMb=statTable->deallocated_elements*8.0/1048576;
-  float tszMb=statTable->max_tab_size*8.0/1048576;
+  double elMb=statTable->allocated_elements*8.0/1048576;
+  double deelMb=statTable->deallocated_elements*8.0/1048576;
+  double tszMb=statTable->max_tab_size*8.0/1048576;
   cout << "======= escript Mem Stats ===========================" << endl;
   cout << "Total Num requests:             " << statTable->requests << endl;
   cout << "Total Num releases:             " << statTable->frees << endl;
